Styling improvements; delete redundancies

+++ src/Novell.Directory.Ldap.NETStandard/Asn1/LBEREncoder.cs
@@ -30,7 +30,6 @@
 // (C) 2003 Novell, Inc (http://www.novell.com)
 //
 
-using System;
 using System.IO;
 
 namespace Novell.Directory.Ldap.Asn1
@@ -78,7 +77,7 @@ namespace Novell.Directory.Ldap.Asn1
             outRenamed.WriteByte(0x01);
 
             /* Encode the boolean content*/
-            outRenamed.WriteByte((byte)(b.BooleanValue() ? (byte)SupportClass.Identity(0xff) : (byte)0x00));
+            outRenamed.WriteByte(b.BooleanValue() ? (byte)SupportClass.Identity(0xff) : (byte)0x00);
         }
 
         /// <summary>
@@ -102,12 +101,12 @@ namespace Novell.Directory.Ldap.Asn1
             }
 
             Encode(n.GetIdentifier(), outRenamed);
-            outRenamed.WriteByte((byte)len); // Length
+            outRenamed.WriteByte(len); // Length
             for (var i = len - 1; i >= 0; i--)
 
                 // Content
             {
-                outRenamed.WriteByte((byte)octets[i]);
+                outRenamed.WriteByte(octets[i]);
             }
         }
