+++ src/Novell.Directory.Ldap.NETStandard/SupportClass.cs
@@ -94,7 +94,7 @@ namespace Novell.Directory.Ldap
 
             for (var i = start; i < start + bytesRead; i++)
             {
-                target[i] = (byte)receiver[i];
+                target[i] = receiver[i];
             }
 
             return bytesRead;
